新增(API)：支持查询条件

+++ Aries-Navigation-Api/aries-navigation-api/src/main/java/com/kuretru/web/aries/service/WebSiteService.java
@@ -34,6 +34,6 @@ public interface WebSiteService extends BaseService<WebSiteDTO, WebSiteQuery> {
      * @param categoryId 指定CategoryID
      * @return 最大的序列编号
      */
-    short getMaxSequence(UUID categoryId);
+    int getMaxSequence(UUID categoryId);
 
 }
